+++ model-impl/src/main/java/com/taxiservice/model/repository/PhoneNumberRepository.java
@@ -3,5 +3,5 @@ package com.taxiservice.model.repository;
 import com.taxiservice.model.entity.PhoneNumber;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PhoneNumberRepository  extends CrudRepository<PhoneNumber, Integer>{
+public interface PhoneNumberRepository  extends CrudRepository<PhoneNumber, Long>{
 }
