+++ model-impl/src/main/java/com/taxiservice/model/repository/TaxiDriverRepository.java
@@ -4,5 +4,5 @@ package com.taxiservice.model.repository;
 import com.taxiservice.model.entity.TaxiDriver;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TaxiDriverRepository extends CrudRepository<TaxiDriver, Integer> {
+public interface TaxiDriverRepository extends CrudRepository<TaxiDriver, Long> {
 }
